refactor getEventArgs
