+++ common/src/index.ts
@@ -5,3 +5,4 @@ export * from './ipfs'
 export * from './keypair'
 export * from './tally'
 export * from './utils'
+export * from './event'
