+++ contracts/utils/contracts.ts
@@ -1,4 +1,5 @@
-import { Contract, TransactionResponse } from 'ethers'
+import { TransactionResponse } from 'ethers'
+import { getEventArg } from '@clrfund/common'
 
 export async function getGasUsage(
   transaction: TransactionResponse
@@ -15,27 +16,4 @@ export async function getTxFee(
   return receipt ? BigInt(receipt.gasUsed) * BigInt(receipt.gasPrice) : 0n
 }
 
-export async function getEventArg(
-  transaction: TransactionResponse,
-  contract: Contract,
-  eventName: string,
-  argumentName: string
-): Promise<any> {
-  const contractAddress = await contract.getAddress()
-  // eslint-disable-line @typescript-eslint/no-explicit-any
-  const receipt = await transaction.wait()
-  for (const log of receipt?.logs || []) {
-    if (log.address != contractAddress) {
-      continue
-    }
-
-    const event = contract.interface.parseLog({
-      data: log.data,
-      topics: [...log.topics],
-    })
-    if (event && event.name === eventName) {
-      return event.args[argumentName]
-    }
-  }
-  throw new Error('Event not found')
-}
+export { getEventArg }
